deeplearning-assignment/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy==1.26.0
opencv-python-headless==4.8.1.78
Pillow==10.0.1
tensorflow-cpu==2.15.0
gunicorn==21.2.0